requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.35.1
black==25.9.0
boto3==1.40.41
botocore==1.40.41
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
isodate==0.7.2
jmespath==1.0.1
loguru==0.7.3
msrest==0.7.1
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyMuPDF==1.26.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
